tests/sql.py: add a "partitions" table to the database
In a few cases, it suffices to partition the dimension when we are
looking for similacra (most recently in Proposition 9, but elsewhere
too). Using admissible_lorentz_ranks() for this fails however if the
target cone is itself a sum of Lorentz cones, because then obviously
it's rank is admissible, but we don't know how _many_ cones had that
same rank. Was it one? Or more?

For this reason it will be useful to have the partitions themselves
cached. If only one partition gives the target rank, then there is
only one sum-of-Lorentz cone with that rank, namely the one whose
similacra you were trying to find.