tests: fix all typing issues (mypy)
Most of these are just fixing type signatures, but in a few places, we
have to avoid mangling an input argument on-the-fly to an output
argument of another type. This _probably_ doesn't slow things down
too much.

Oh, and it looks like typing for _merge_factors() is hopeless because
we can't narrow the types of "a" and "b" down with a "match"
statement.  As a result, at the end, mypy complains about the
(impossible) possibility of adding an int to a tuple.