Try to de-confuse the proof of Theorem 4
By splitting this into "all Lorentz" and "not all Lorentz" cases, we
avoid the suspicious footnote. It also helps to cite Lemma 4 at the
very beginning to rule out L(n+k) factors right away. The remaining
contradictions are then not even contradictions; they just rule out
the possibility that L(n) is not a factor of J.