Combine rows in the n/dim(K) table in Theorem 5
Some of the values of dim(K) were repeated, so it makes sense to
combine the "n" for which that happens. Subsequently the \set{}
wrapper around the dim(K) values was dropped to make it consistent
with the "n" column; this looked better than adding \set{} around the
"n" values. Finally, the table is now left-aligned; again, this just
looks better now that there are multiple "n" values.